+++ NEWS.md
@@ -10,6 +10,7 @@ This may have been addressed at 64a308ba50b9119108e8946737460de5997b805b by addi
 
 # vcfR 1.5.0.9000
 Released on CRAN 201X-XX-XX.
+* `vcfR2DNAbin()` can include indels and maintains alignment.
 * `write.vcf()` now handles tilde expansion.
 * `rePOS()` attempts top create a non-overlapping coordinate system from POS and CHROM.
 * `vcfR2DNAbin()` manages the asterisk allele.